+++ utils/json-utils/src/main/scala/pl/touk/nussknacker/engine/json/JsonSchemaBuilder.scala
@@ -7,15 +7,15 @@ import org.json.JSONObject
 object JsonSchemaBuilder {
 
   def parseSchema(rawJsonSchema: String): Schema =
-    parseSchema(rawJsonSchema, useDefaults = true, nullableSupport = true)
+    parseSchema(rawJsonSchema, useDefaults = true)
 
-  def parseSchema(rawJsonSchema: String, useDefaults: Boolean, nullableSupport: Boolean): Schema = {
+  //TODO: parsing schema should be consistent with [io.confluent.kafka.schemaregistry.json.JsonSchema.rawSchema]
+  def parseSchema(rawJsonSchema: String, useDefaults: Boolean): Schema = {
     val rawSchema: JSONObject = new JSONObject(rawJsonSchema)
 
     SchemaLoader
       .builder()
       .useDefaults(useDefaults)
-      .nullableSupport(nullableSupport)
       .schemaJson(rawSchema)
       .draftV7Support()
       .build()